refactor(test): reorganize sandbox tests into contract/ptc/providers modules
