+++ tests/integration/sandbox/contract/test_data_persistence.py
@@ -1,0 +1,46 @@
+"""Contract tests for data persistence across stop/start cycles."""
+
+from __future__ import annotations
+
+import pytest
+
+pytestmark = [pytest.mark.integration, pytest.mark.asyncio]
+
+
+class TestDataPersistenceAcrossCycles:
+    """Verify files survive stop -> start cycles (provider contract)."""
+
+    async def test_files_persist_across_stop_start(self, sandbox_runtime, timed):
+        wd = await sandbox_runtime.fetch_working_dir()
+        await sandbox_runtime.upload_file(b"persistent data", f"{wd}/persist.txt")
+
+        async with timed("persistence", "stop_start_persist"):
+            await sandbox_runtime.stop()
+            await sandbox_runtime.start()
+
+        content = await sandbox_runtime.download_file(f"{wd}/persist.txt")
+        assert content == b"persistent data"
+
+    async def test_code_run_output_files_persist(self, sandbox_runtime, timed):
+        wd = await sandbox_runtime.fetch_working_dir()
+        await sandbox_runtime.code_run(
+            f"with open('{wd}/code_file.txt', 'w') as f: f.write('from code_run')"
+        )
+
+        async with timed("persistence", "stop_start_code_persist"):
+            await sandbox_runtime.stop()
+            await sandbox_runtime.start()
+
+        content = await sandbox_runtime.download_file(f"{wd}/code_file.txt")
+        assert content == b"from code_run"
+
+    async def test_multiple_stop_start_cycles(self, sandbox_runtime, timed):
+        wd = await sandbox_runtime.fetch_working_dir()
+        await sandbox_runtime.upload_file(b"cycle_data", f"{wd}/cycle.txt")
+
+        async with timed("persistence", "multiple_cycles"):
+            for i in range(3):
+                await sandbox_runtime.stop()
+                await sandbox_runtime.start()
+                content = await sandbox_runtime.download_file(f"{wd}/cycle.txt")
+                assert content == b"cycle_data", f"Failed on cycle {i}"
