+++ tests/integration/sandbox/ptc/test_path_normalization.py
@@ -1,0 +1,56 @@
+from __future__ import annotations
+
+import pytest
+
+pytestmark = pytest.mark.integration
+
+
+class TestPathNormalization:
+    """PTCSandbox path normalization and validation (sync -- no asyncio mark)."""
+
+    def test_normalize_dot(self, sandbox):
+        wd = sandbox.config.filesystem.working_directory
+        assert sandbox.normalize_path(".") == wd
+
+    def test_normalize_empty(self, sandbox):
+        wd = sandbox.config.filesystem.working_directory
+        assert sandbox.normalize_path("") == wd
+
+    def test_normalize_slash(self, sandbox):
+        wd = sandbox.config.filesystem.working_directory
+        assert sandbox.normalize_path("/") == wd
+
+    def test_normalize_relative(self, sandbox):
+        wd = sandbox.config.filesystem.working_directory
+        assert sandbox.normalize_path("data/file.txt") == f"{wd}/data/file.txt"
+
+    def test_normalize_virtual_absolute(self, sandbox):
+        wd = sandbox.config.filesystem.working_directory
+        assert sandbox.normalize_path("/results/out.txt") == f"{wd}/results/out.txt"
+
+    def test_normalize_already_absolute(self, sandbox):
+        wd = sandbox.config.filesystem.working_directory
+        assert sandbox.normalize_path(f"{wd}/data/x.txt") == f"{wd}/data/x.txt"
+
+    def test_normalize_tmp(self, sandbox):
+        assert sandbox.normalize_path("/tmp/file.txt") == "/tmp/file.txt"
+
+    def test_virtualize_path(self, sandbox):
+        wd = sandbox.config.filesystem.working_directory
+        assert sandbox.virtualize_path(f"{wd}/data/x.txt") == "/data/x.txt"
+        assert sandbox.virtualize_path(wd) == "/"
+        assert sandbox.virtualize_path("/tmp/x.txt") == "/tmp/x.txt"
+
+    def test_validate_allowed(self, sandbox):
+        wd = sandbox.config.filesystem.working_directory
+        assert sandbox.validate_path(f"{wd}/data/x.txt") is True
+        assert sandbox.validate_path("/tmp/x.txt") is True
+
+    def test_validate_denied(self, sandbox):
+        wd = sandbox.config.filesystem.working_directory
+        # Add a denied directory to test denial
+        sandbox.config.filesystem.denied_directories = [f"{wd}/_internal"]
+        assert sandbox.validate_path(f"{wd}/_internal/secret.txt") is False
+        assert sandbox.validate_path(f"{wd}/data/ok.txt") is True
+        # Restore
+        sandbox.config.filesystem.denied_directories = []
